+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [2.4.0] - 2021-09-09
 ### Added
+- process_dbs_tbls:  Process a list of databases and tables.
 - process_mail:  Add data to mail instance and send mail.
 
 ### Changed